+++ packages/good/test/multi_scene_test.dart
@@ -23,11 +23,7 @@ late Game run;
 mixin _Marked on Component {
   final mark = Field.uint16(3);
 
-  @override
-  void describeType(ComponentDescriptor component) {
-    super.describeType(component);
-    component.has<_Marked>();
-  }
+  final markedType = Component.type<_Marked>();
 }
 
 class _Unit extends EntityStruct with _Marked {}